refactor(api): move usecase to src/team
